backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
fpdf2
pybind11